pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aqueduct-core"
version = "1.0.0a0"
description = "Declarative, self-healing Apache Spark blueprints."
readme = "README.md"
authors = [
    {name = "Sadig Akhund", email = "sadigaxund@gmail.com"},
]
license = {file = "LICENSE"}
requires-python = ">=3.11"
keywords = ["spark", "blueprint", "pipeline", "ETL", "data-engineering", "llm", "self-healing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "duckdb>=0.10.0",
    "sqlglot>=23.0.0",
    "ruamel.yaml>=0.18.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
spark = ["pyspark>=3.5.0", "delta-spark>=3.0.0"]
all = ["aqueduct-core[spark]"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/sadigaxund/Aqueduct"
Repository = "https://github.com/sadigaxund/Aqueduct"
Issues = "https://github.com/sadigaxund/Aqueduct/issues"
Changelog = "https://github.com/sadigaxund/Aqueduct/blob/main/CHANGELOG.md"

[project.scripts]
aqueduct = "aqueduct.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["aqueduct"]

[tool.hatch.build.targets.sdist]
include = [
    "/aqueduct",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
    "/CHANGELOG.md",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts   = "--cov=aqueduct --cov-report=term-missing"
testpaths = ["tests"]
cache_dir = "/tmp/pytest_cache_aqueduct"

[tool.coverage.run]
data_file = "/tmp/.aqueduct_coverage"
